Redefine "Marshal" to take only a value, not a type
The cty JSON and msgpack packages both use this to do some trickery to
preserve in the saved object type information for anything that is
unconstrained in the given type constraint, so that it can be recovered
exactly by Unmarshal.

We won't do that here, since it isn't yet a goal for this package to be
able to preserve exact cty type information in a round-trip through YAML,
and if we _do_ decide to try that later it'll be better achieved through
custom YAML type tags rather than in-band type information as we do in
the other two serializations.