Implicit number recognition pattern now matches YAML specifications
The regular expression pattern used to infer when an untagged scalar was intended to be a number didn't quite match the rules in the YAML specifications. In particular, it was allowing underscores when the YAML spec doesn't and was then incorrectly handling a string containing only a single underscore without any digits.

The number pattern now matches what YAML requires, which means underscores are not allowed and a sign prefix is only allowed for decimal numbers. Untagged scalars not matching this pattern or one of the other implied type recognition patterns will be returned as strings so that the caller can then parse them further if desired.