Allow a !!merge key to be used with a sequence of mappings
Previously we had some basic support for !!merge-tagged values, which are
usually just implied by using the scalar << in mapping key position, but
it was implemented in a quirky way and only worked for direct mappings an
not for sequences of mappings.

This is now implemented in a way that seems more consistent with other YAML
implementations, where this tag is treated as a special case only in the
mapping key position and the value can be a YAML sequence.

This intentionally slightly violates the spec for this tag by letting a
merged mapping overwrite a key already present in the mapping. Despite the
spec saying that only keys that are not already present should be included,
many existing implementations follow the behavior implemented here and it
seems like a more intuitive interpretation given how duplicate mapping keys
are normally handled in YAML.

Overall the spec for !!merge is terse, vague, and incomplete. This was
based on what seems to be the prevailing understanding of it in existing
implementations, which is hopefully good enough for interoperability.